Rename Assignment5.py to HW5.py
Hello world Python program that will print my name, the date, and lists of various element sizes will be created, sorted, and verified for ascending order. The amount of time it takes to sort each list will be collected into a new list. After all lists are run through, we will plot a graph of the time elapsed vs the elements in a problem size.